+++ test/tactIR.spec.ts
@@ -1,4 +1,4 @@
-import { run } from "../src/driver";
+import { Runner } from "../src/driver";
 import { TAP, processTactFiles, GOOD_DIR, resetIds } from "./testUtil";
 import path from "path";
 
@@ -9,22 +9,24 @@ processTactFiles(GOOD_DIR, (file) => {
   describe(`Testing CFG dump for ${contractName}`, () => {
     it(`should produce correct CFG JSON output for ${contractName}`, async () => {
       resetIds();
-      await run(filePath, {
+      const runner = await Runner.make(filePath, {
         dumpCfg: "json",
         dumpCfgStdlib: false,
         dumpCfgOutput: GOOD_DIR,
         quiet: true,
       });
+      runner.run();
       await TAP.from(nameBase, "json", "cfg.json").run();
     });
     it(`should produce correct CFG DOT output for ${contractName}`, async () => {
       resetIds();
-      await run(filePath, {
+      const runner = await Runner.make(filePath, {
         dumpCfg: "dot",
         dumpCfgStdlib: false,
         dumpCfgOutput: GOOD_DIR,
         quiet: true,
       });
+      runner.run();
       await TAP.from(nameBase, "dot", "cfg.dot").run();
     });
   });